+++ .expo/settings.json
@@ -1,8 +1,6 @@
 {
-  "hostType": "lan",
-  "lanType": "ip",
-  "dev": true,
-  "minify": false,
-  "urlRandomness": null,
-  "https": false
+"dev": true,
+"minify": false,
+"urlRandomness": null,
+"https": false
 }
